session/handler_item_stack_request.go: Split up anvils, crafting, and beacons.
