+++ server/session/handler_item_stack_request.go
@@ -2,19 +2,12 @@ package session
 
 import (
 	"fmt"
-	"github.com/df-mc/dragonfly/server/block"
-	"github.com/df-mc/dragonfly/server/entity/effect"
 	"github.com/df-mc/dragonfly/server/event"
 	"github.com/df-mc/dragonfly/server/item"
-	"github.com/df-mc/dragonfly/server/item/creative"
 	"github.com/df-mc/dragonfly/server/item/inventory"
-	"github.com/df-mc/dragonfly/server/item/recipe"
-	"github.com/df-mc/dragonfly/server/world/sound"
 	"github.com/sandertv/gophertunnel/minecraft/protocol"
 	"github.com/sandertv/gophertunnel/minecraft/protocol/packet"
-	"golang.org/x/exp/slices"
 	"math"
-	"math/rand"
 	"time"
 )
 
@@ -194,187 +187,6 @@ func call(ctx *event.Context, slot int, it item.Stack, f func(ctx *event.Context
 	return nil
 }
 
-// handleCraft handles the CraftRecipe request action.
-func (h *ItemStackRequestHandler) handleCraft(a *protocol.CraftRecipeStackRequestAction, s *Session) error {
-	craft, ok := s.recipes[a.RecipeNetworkID]
-	if !ok {
-		return fmt.Errorf("recipe with network id %v does not exist", a.RecipeNetworkID)
-	}
-	_, shaped := craft.(recipe.Shaped)
-	_, shapeless := craft.(recipe.Shapeless)
-	if !shaped && !shapeless {
-		return fmt.Errorf("recipe with network id %v is not a shaped or shapeless recipe", a.RecipeNetworkID)
-	}
-
-	size := s.craftingSize()
-	offset := s.craftingOffset()
-	consumed := make([]bool, size)
-	for _, expected := range craft.Input() {
-		var processed bool
-		for slot := offset; slot < offset+size; slot++ {
-			if consumed[slot-offset] {
-				// We've already consumed this slot, skip it.
-				continue
-			}
-			has, _ := s.ui.Item(int(slot))
-			_, variants := expected.Value("variants")
-			if has.Empty() != expected.Empty() || has.Count() < expected.Count() {
-				// We can't process this item, as it's not a part of the recipe.
-				continue
-			}
-			if !variants && !has.Comparable(expected) {
-				// Not the same item without accounting for variants.
-				continue
-			}
-			if variants {
-				nameOne, _ := has.Item().EncodeItem()
-				nameTwo, _ := expected.Item().EncodeItem()
-				if nameOne != nameTwo {
-					// Not the same item even when accounting for variants.
-					continue
-				}
-			}
-			processed, consumed[slot-offset] = true, true
-			st := has.Grow(-expected.Count())
-			h.setItemInSlot(protocol.StackRequestSlotInfo{
-				ContainerID:    containerCraftingGrid,
-				Slot:           byte(slot),
-				StackNetworkID: item_id(st),
-			}, st, s)
-			break
-		}
-		if !processed {
-			return fmt.Errorf("recipe %v: could not consume expected item: %v", a.RecipeNetworkID, expected)
-		}
-	}
-
-	output := craft.Output()
-	h.setItemInSlot(protocol.StackRequestSlotInfo{
-		ContainerID:    containerCraftingGrid,
-		Slot:           craftingGridResult,
-		StackNetworkID: item_id(output[0]),
-	}, output[0], s)
-	return nil
-}
-
-// handleAutoCraft handles the AutoCraftRecipe request action.
-func (h *ItemStackRequestHandler) handleAutoCraft(a *protocol.AutoCraftRecipeStackRequestAction, s *Session) error {
-	craft, ok := s.recipes[a.RecipeNetworkID]
-	if !ok {
-		return fmt.Errorf("recipe with network id %v does not exist", a.RecipeNetworkID)
-	}
-	_, shaped := craft.(recipe.Shaped)
-	_, shapeless := craft.(recipe.Shapeless)
-	if !shaped && !shapeless {
-		return fmt.Errorf("recipe with network id %v is not a shaped or shapeless recipe", a.RecipeNetworkID)
-	}
-
-	input := make([]item.Stack, 0, len(craft.Input()))
-	for _, i := range craft.Input() {
-		input = append(input, i.Grow(i.Count()*(int(a.TimesCrafted)-1)))
-	}
-
-	expectancies := make([]item.Stack, 0, len(input))
-	for _, i := range input {
-		if i.Empty() {
-			// We don't actually need this item - it's empty, so avoid putting it in our expectancies.
-			continue
-		}
-
-		_, variants := i.Value("variants")
-		if ind := slices.IndexFunc(expectancies, func(st item.Stack) bool {
-			if variants {
-				nameOne, _ := st.Item().EncodeItem()
-				nameTwo, _ := i.Item().EncodeItem()
-				return nameOne == nameTwo
-			}
-			return st.Comparable(i)
-		}); ind >= 0 {
-			i = i.Grow(expectancies[ind].Count())
-			expectancies = slices.Delete(expectancies, ind, ind+1)
-		}
-		expectancies = append(expectancies, i)
-	}
-
-	for _, expected := range expectancies {
-		_, variants := expected.Value("variants")
-		for id, inv := range map[byte]*inventory.Inventory{containerCraftingGrid: s.ui, containerFullInventory: s.inv} {
-			for slot, has := range inv.Slots() {
-				if has.Empty() {
-					// We don't have this item, skip it.
-					continue
-				}
-				if !variants && !has.Comparable(expected) {
-					// Not the same item without accounting for variants.
-					continue
-				}
-				if variants {
-					nameOne, _ := has.Item().EncodeItem()
-					nameTwo, _ := expected.Item().EncodeItem()
-					if nameOne != nameTwo {
-						// Not the same item even when accounting for variants.
-						continue
-					}
-				}
-
-				remaining, removal := expected.Count(), has.Count()
-				if remaining < removal {
-					removal = remaining
-				}
-
-				expected, has = expected.Grow(-removal), has.Grow(-removal)
-				h.setItemInSlot(protocol.StackRequestSlotInfo{
-					ContainerID:    id,
-					Slot:           byte(slot),
-					StackNetworkID: item_id(has),
-				}, has, s)
-				if expected.Empty() {
-					// Consumed this item, so go to the next one.
-					break
-				}
-			}
-			if expected.Empty() {
-				// Consumed this item, so go to the next one.
-				break
-			}
-		}
-		if !expected.Empty() {
-			return fmt.Errorf("recipe %v: could not consume expected item: %v", a.RecipeNetworkID, expected)
-		}
-	}
-
-	output := make([]item.Stack, 0, len(craft.Output()))
-	for _, o := range craft.Output() {
-		output = append(output, o.Grow(o.Count()*(int(a.TimesCrafted)-1)))
-	}
-	h.setItemInSlot(protocol.StackRequestSlotInfo{
-		ContainerID:    containerCraftingGrid,
-		Slot:           craftingGridResult,
-		StackNetworkID: item_id(output[0]),
-	}, output[0], s)
-	return nil
-}
-
-// handleCreativeCraft handles the CreativeCraft request action.
-func (h *ItemStackRequestHandler) handleCreativeCraft(a *protocol.CraftCreativeStackRequestAction, s *Session) error {
-	if !s.c.GameMode().CreativeInventory() {
-		return fmt.Errorf("can only craft creative items in gamemode creative/spectator")
-	}
-	index := a.CreativeItemNetworkID - 1
-	if int(index) >= len(creative.Items()) {
-		return fmt.Errorf("creative item with network ID %v does not exist", index)
-	}
-	it := creative.Items()[index]
-	it = it.Grow(it.MaxCount() - 1)
-
-	h.setItemInSlot(protocol.StackRequestSlotInfo{
-		ContainerID:    containerOutput,
-		Slot:           50,
-		StackNetworkID: item_id(it),
-	}, it, s)
-	return nil
-}
-
 // handleDestroy handles the destroying of an item by moving it into the creative inventory.
 func (h *ItemStackRequestHandler) handleDestroy(a *protocol.DestroyStackRequestAction, s *Session) error {
 	if h.ignoreDestroy {
@@ -416,54 +228,6 @@ func (h *ItemStackRequestHandler) handleDrop(a *protocol.DropStackRequestAction,
 	return nil
 }
 
-// handleBeaconPayment handles the selection of effects in a beacon and the removal of the item used to pay
-// for those effects.
-func (h *ItemStackRequestHandler) handleBeaconPayment(a *protocol.BeaconPaymentStackRequestAction, s *Session) error {
-	slot := protocol.StackRequestSlotInfo{
-		ContainerID: containerBeacon,
-		Slot:        0x1b,
-	}
-	// First check if there actually is a beacon opened.
-	if !s.containerOpened.Load() {
-		return fmt.Errorf("no beacon container opened")
-	}
-	pos := s.openedPos.Load()
-	beacon, ok := s.c.World().Block(pos).(block.Beacon)
-	if !ok {
-		return fmt.Errorf("no beacon container opened")
-	}
-
-	// Check if the item present in the beacon slot is valid.
-	payment, _ := h.itemInSlot(slot, s)
-	if payable, ok := payment.Item().(item.BeaconPayment); !ok || !payable.PayableForBeacon() {
-		return fmt.Errorf("item %#v in beacon slot cannot be used as payment", payment)
-	}
-
-	// Check if the effects are valid and allowed for the beacon's level.
-	if !h.validBeaconEffect(a.PrimaryEffect, beacon) {
-		return fmt.Errorf("primary effect selected is not allowed: %v for level %v", a.PrimaryEffect, beacon.Level())
-	} else if !h.validBeaconEffect(a.SecondaryEffect, beacon) || (beacon.Level() < 4 && a.SecondaryEffect != 0) {
-		return fmt.Errorf("secondary effect selected is not allowed: %v for level %v", a.SecondaryEffect, beacon.Level())
-	}
-
-	primary, pOk := effect.ByID(int(a.PrimaryEffect))
-	secondary, sOk := effect.ByID(int(a.SecondaryEffect))
-	if pOk {
-		beacon.Primary = primary.(effect.LastingType)
-	}
-	if sOk {
-		beacon.Secondary = secondary.(effect.LastingType)
-	}
-	s.c.World().SetBlock(pos, beacon, nil)
-
-	// The client will send a Destroy action after this action, but we can't rely on that because the client
-	// could just not send it.
-	// We just ignore the next Destroy action and set the item to air here.
-	h.setItemInSlot(slot, item.NewStack(block.Air{}, 0), s)
-	h.ignoreDestroy = true
-	return nil
-}
-
 // handleMineBlock handles the action associated with a block being mined by the player. This seems to be a workaround
 // by Mojang to deal with the durability changes client-side.
 func (h *ItemStackRequestHandler) handleMineBlock(a *protocol.MineBlockStackRequestAction, s *Session) error {
@@ -479,236 +243,7 @@ func (h *ItemStackRequestHandler) handleMineBlock(a *protocol.MineBlockStackRequ
 	// Update the slots through ItemStackResponses, don't actually do anything special with this action.
 	i, _ := h.itemInSlot(slot, s)
 	h.setItemInSlot(slot, i, s)
-
 	return nil
-}
-
-// handleCraftRecipeOptional ...
-func (h *ItemStackRequestHandler) handleCraftRecipeOptional(a *protocol.CraftRecipeOptionalStackRequestAction, s *Session, filterStrings []string) error {
-	// First check if there actually is an anvil opened.
-	if !s.containerOpened.Load() {
-		return fmt.Errorf("no anvil container opened")
-	}
-	pos := s.openedPos.Load()
-	w := s.c.World()
-	anvil, ok := w.Block(pos).(block.Anvil)
-	if !ok {
-		return fmt.Errorf("no anvil container opened")
-	}
-	if len(filterStrings) < int(a.FilterStringIndex) {
-		// Invalid filter string index.
-		return nil
-	}
-
-	first, _ := h.itemInSlot(protocol.StackRequestSlotInfo{
-		ContainerID: containerAnvilInput,
-		Slot:        1,
-	}, s)
-	if first.Empty() {
-		// First anvil slot is empty, can't result in anything.
-		return nil
-	}
-	result := first
-
-	second, _ := h.itemInSlot(protocol.StackRequestSlotInfo{
-		ContainerID: containerAnvilMaterial,
-		Slot:        0x2,
-	}, s)
-
-	cost := first.RepairCost()
-	if !second.Empty() {
-		cost += second.RepairCost()
-	}
-
-	var repairCount int
-	if !second.Empty() {
-		if repairable, ok := first.Item().(item.Repairable); ok && repairable.RepairableBy(second) {
-			d := min(first.MaxDurability()-first.Durability(), first.MaxDurability()/4)
-			if d <= 0 {
-				return nil
-			}
-
-			for ; d > 0 && repairCount < second.Count(); repairCount, d = repairCount+1, min(result.MaxDurability()-result.Durability(), result.MaxDurability()/4) {
-				result = result.WithDurability(result.Durability() + d)
-				cost++
-			}
-		} else {
-			_, book := second.Item().(item.EnchantedBook)
-			_, durable := first.Item().(item.Durable)
-
-			enchant := book && len(second.Enchantments()) > 0
-			if !enchant && (first.Item() != second.Item() || !durable) {
-				return nil
-			}
-			if durable && !enchant {
-				d := first.MaxDurability() - (first.Durability() + (second.Durability() + first.MaxDurability()*12/100))
-				if d < 0 {
-					d = 0
-				}
-				if d < first.MaxDurability()-first.Durability() {
-					result = result.WithDurability(d)
-					cost += 2
-				}
-			}
-
-			var hasCompatible, hasIncompatible bool
-			for _, e := range second.Enchantments() {
-				t := e.Type()
-
-				var firstLevel int
-				if firstEnchant, ok := first.Enchantment(t); ok {
-					firstLevel = firstEnchant.Level()
-				}
-
-				resultLevel := max(firstLevel, e.Level())
-				if firstLevel == e.Level() {
-					resultLevel = firstLevel + 1
-				}
-
-				compatible := t.CompatibleWithItem(first.Item())
-				if _, ok := first.Item().(item.EnchantedBook); ok {
-					compatible = true
-				}
-
-				for _, e2 := range first.Enchantments() {
-					if t != e2.Type() && !t.CompatibleWithOther(e2.Type()) {
-						compatible = false
-						cost++
-					}
-				}
-
-				if !compatible {
-					hasIncompatible = true
-					continue
-				}
-				hasCompatible = true
-
-				if resultLevel > t.MaxLevel() {
-					resultLevel = t.MaxLevel()
-				}
-				rarityCost := t.Rarity().ApplyCost
-				if enchant {
-					rarityCost = max(1, rarityCost/2)
-				}
-
-				result = result.WithEnchantments(item.NewEnchantment(t, resultLevel))
-				cost += rarityCost * resultLevel
-				if first.Count() > 1 {
-					cost = 40
-				}
-			}
-			if hasIncompatible && !hasCompatible {
-				// We have no compatible enchantments, but we have incompatible ones.
-				return nil
-			}
-		}
-	}
-
-	newName := filterStrings[int(a.FilterStringIndex)]
-	existingName := item.DisplayName(first.Item(), s.c.Locale())
-	if customName := first.CustomName(); len(customName) > 0 {
-		existingName = customName
-	}
-	if existingName != newName {
-		result = result.WithCustomName(newName)
-		cost += 1
-	}
-
-	if cost == 0 {
-		// No action was performed.
-		return nil
-	}
-
-	c := s.c.GameMode().CreativeInventory()
-	if cost >= 40 && !c {
-		// Impossible repair/rename.
-		return nil
-	}
-
-	if !result.Empty() {
-		i := result.RepairCost()
-		if !second.Empty() && i < second.RepairCost() {
-			i = second.RepairCost()
-		}
-		if cost != 1 {
-			i = i*2 + 1
-		}
-		result = result.WithRepairCost(i)
-	}
-
-	level := s.c.ExperienceLevel()
-	if level < cost && !c {
-		// Not enough experience.
-		return nil
-	} else if !c {
-		s.c.SetExperienceLevel(level - cost)
-	}
-
-	if !c && rand.Float64() < 0.12 {
-		damaged := anvil.Break()
-		if _, ok := damaged.(block.Air); ok {
-			w.PlaySound(pos.Vec3Centre(), sound.AnvilBreak{})
-		} else {
-			w.PlaySound(pos.Vec3Centre(), sound.AnvilUse{})
-		}
-		defer w.SetBlock(pos, damaged, nil)
-	} else {
-		w.PlaySound(pos.Vec3Centre(), sound.AnvilUse{})
-	}
-
-	h.setItemInSlot(protocol.StackRequestSlotInfo{
-		ContainerID: containerAnvilInput,
-		Slot:        1,
-	}, item.Stack{}, s)
-	if repairCount > 0 {
-		h.setItemInSlot(protocol.StackRequestSlotInfo{
-			ContainerID: containerAnvilMaterial,
-			Slot:        2,
-		}, second.Grow(-repairCount), s)
-	} else {
-		h.setItemInSlot(protocol.StackRequestSlotInfo{
-			ContainerID: containerAnvilMaterial,
-			Slot:        2,
-		}, item.Stack{}, s)
-	}
-	h.setItemInSlot(protocol.StackRequestSlotInfo{
-		ContainerID: containerOutput,
-		Slot:        50,
-	}, result, s)
-	return nil
-}
-
-// max returns the max of two integers.
-func max(x, y int) int {
-	if x > y {
-		return x
-	}
-	return y
-}
-
-// min returns the min of two integers.
-func min(x, y int) int {
-	if x > y {
-		return y
-	}
-	return x
-}
-
-// validBeaconEffect checks if the ID passed is a valid beacon effect.
-func (h *ItemStackRequestHandler) validBeaconEffect(id int32, beacon block.Beacon) bool {
-	switch id {
-	case 1, 3:
-		return beacon.Level() >= 1
-	case 8, 11:
-		return beacon.Level() >= 2
-	case 5:
-		return beacon.Level() >= 3
-	case 10:
-		return beacon.Level() >= 4
-	case 0:
-		return true
-	}
-	return false
 }
 
 // verifySlots verifies a list of slots passed.
